LayoutTests/imported/w3c/web-platform-tests/xhr/send-conditional-cors-expected.txt

Blocked access to external URL http://www1.localhost:8800/XMLHttpRequest/resources/conditional.py?tag=0.45864598185251126
CONSOLE MESSAGE: line 37: XMLHttpRequest cannot load http://www1.localhost:8800/XMLHttpRequest/resources/conditional.py?tag=0.45864598185251126 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/XMLHttpRequest/resources/conditional.py?tag=0.3336103787755589&cors=yes
CONSOLE MESSAGE: line 37: XMLHttpRequest cannot load http://www1.localhost:8800/XMLHttpRequest/resources/conditional.py?tag=0.3336103787755589&cors=yes due to access control checks.

Harness Error (TIMEOUT), message = null

TIMEOUT 304 without appropriate CORS header Test timed out
TIMEOUT 304 with appropriate CORS header Test timed out
